adaptive dicing implemented & quaternion fix

+++ base/framebuffer.hpp
@@ -6,11 +6,11 @@
 #include "./vertex.hpp"
 #include <opencv2/opencv.hpp>
 
-//data needed per sample
+// data needed per sample
 struct Sample
 {
     float depth;
-    //in rgba format
+    // in rgba format
     gl::vec4 color;
 
     void setDepth(float d)
@@ -26,7 +26,7 @@ struct Sample
     Sample()
     {
         this->depth = 1.0;
-        this->color = gl::vec4(gl::vec3(0.0f),1.0f);
+        this->color = gl::vec4(gl::vec3(0.0f), 1.0f);
     }
 
     Sample(float depth, gl::vec4 color)
@@ -47,15 +47,15 @@ class FrameBuffer
 public:
     FrameBuffer(uint width, uint height, uint m = 1, uint n = 1);
     void setPixelColor(uint u, uint v, gl::vec3 color);
-    //use sample buffer to set pixel color
+    // use sample buffer to set pixel color
     void setPixelColorFromBuffer();
-    //set sample number
+    // set sample number
     void setSampleNumber(uint m, uint n);
-    //initialize all sample buffers
+    // initialize all sample buffers
     void initBuffers();
 
     // given a micropolygon, test it with the framebuffer, update buffers accordingly
-    void updateBufferByMicropolygon(Micropolygon &mp,gl::mat4 mvp);
+    void updateBufferByMicropolygon(Micropolygon &mp, gl::mat4 mvp);
 
     uint getSampleNumber() const
     {
@@ -77,9 +77,10 @@ public:
         return this->height;
     };
 
-    //reset all pixel color to 0.0
-    void clearColor(){
-        //set all pixel color to 0.0
+    // reset all pixel color to 0.0
+    void clearColor()
+    {
+        // set all pixel color to 0.0
         for (uint i = 0; i < width; i++)
         {
             for (uint j = 0; j < height; j++)
@@ -89,17 +90,18 @@ public:
         }
     }
 
-    //reset all samples in the buffer to 0.0
-    void clearBuffer(float bg_depth,gl::vec4 bg_color);
+    // reset all samples in the buffer to 0.0
+    void clearBuffer(float bg_depth, gl::vec4 bg_color);
 
     // the input should be 2 integers
     gl::vec3 getPixelColor(uint u, uint v) const
     {
         return this->_pixel_color[u][v];
     };
-    
+
     // suppose color is vec3 and range 0~1
-    cv::Mat to_cv_mat(int FORMAT=CV_8UC3);
+    // conversion happens if the dest format is 8UC3, which loses precision
+    cv::Mat to_cv_mat(int FORMAT = CV_32FC3);
 
 private:
     // 1st 2nd index for pixel, 3rd index for vector of samples
@@ -115,4 +117,3 @@ private:
     uint width;
     uint height;
 };
-
